Add previous page functionality and optimize movie loading
Implemented functionality to load the previous page of movies and optimize the process of loading movies by adding a check to avoid duplication of movies. Additionally, introduced a correction in UI to hide the previous button on the initial page and show only when there's a previous page to load.