Add movie overview on hover
Implemented a new feature in the movie listing section that displays movie overviews when users hover over a movie. The movie information is populated dynamically from the fetched movie data, and the overview is hidden until the user hovers over the movie element. Made some minor changes to comments and removed an unnecessary comment.